soc: bcm: bcm2835-pm: Add support for 2711.
commit  4d8cac4592a4d080df602d60849067fb613f839b from
https://github.com/raspberrypi/linux.git rpi-5.8.y

Without the actual power management part any more, there's a lot less
to set up for V3D.  We just need to clear the RSTN field for the power
domain, and expose the reset controller for toggling it again.

This is definitely incomplete -- the old ISP and H264 is in the old
bridge, but since we have no consumers of it I've just done the
minimum to get V3D working.

Signed-off-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Meng Li <Meng.Li@windriver.com>