PCI: brcmstb: Add MSI capability
commit  b2d621c220d7931ff3123b54d7c1945c57e05970 from
https://github.com/raspberrypi/linux.git

This commit adds MSI to the Broadcom STB PCIe host controller. It does
not add MSIX since that functionality is not in the HW.  The MSI
controller is physically located within the PCIe block, however, there
is no reason why the MSI controller could not be moved elsewhere in
the future.

Since the internal Brcmstb MSI controller is intertwined with the PCIe
controller, it is not its own platform device but rather part of the
PCIe platform device.

Signed-off-by: Jim Quinlan <jim2101024@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>