usb: xhci: Disable the XHCI 5 second timeout
commit  18f73fc091e63b190b162929885df4e434101263 from
https://github.com/raspberrypi/linux.git rpi-5.8.y

If the VL805 EEPROM has not been programmed then boot will hang for five
seconds. The timeout seems to be arbitrary and is an unecessary
delay on the first boot. Remove the timeout.

This is common code and probably can't be upstreamed unless the timeout
can be configurable somehow or perhaps the XHCI driver can be skipped
on the first boot.
Signed-off-by: Meng Li <Meng.Li@windriver.com>