usb: xhci: Disable the XHCI 5 second timeout
commit  b40771a19a1e0b346d913f385c93db6a34dc16d2 from
https://github.com/raspberrypi/linux.git

If the VL805 EEPROM has not been programmed then boot will hang for five
seconds. The timeout seems to be arbitrary and is an unecessary
delay on the first boot. Remove the timeout.

This is common code and probably can't be upstreamed unless the timeout
can be configurable somehow or perhaps the XHCI driver can be skipped
on the first boot.
Signed-off-by: Meng Li <Meng.Li@windriver.com>