net: xilinx: axiethernet: Fix xxv mac short frame handling
Get rid of custom short packet padding code. The eth_skb_pad() is used
instead, which pads short frames to the 60 byte limit.

The original code was cumbersome and still missing few cases.
It would just run over the packet data to perform padding, which
is wrong. Resulting in DMA beyond the mapped region or copy previous
packet data. It was also padding to 64 bytes instead of 60 (the 4 extra
bytes is FCS added by the core automatically)

As a side-effect it also fixes incorrect FCS reported with 64 byte packet
length.

Signed-off-by: Edgar Lakis <ela@phaseone.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>