drm/vc4: Max resolution of 7680 is conditional on being Pi4
commit  08b1741e20ee77f1959f007fb4b8fd6e5ba1cb5d from
https://github.com/raspberrypi/linux.git rpi-5.8.y

The max resolution had been increased from 2048 to 7680 for all
platforms. This code is common with Pi0-3 which have a max render
target for GL of 2048, therefore the increased resolution has to
be conditional on the platform.
Switch based on whether the bcm2835-v3d node is found, as that is
not present on Pi4. (There is a potential configuration on Pi0-3
with no v3d, but this is very unlikely).

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>