staging: vcsm-cma: Rework to use dma APIs, not CMA
commit  8d43cbe3234f44d6357b8bf08e3339323cee1dab from
https://github.com/raspberrypi/linux.git

Due to a misunderstanding of the DMA mapping APIs, I made
the wrong decision on how to implement this.

Rework to use dma_alloc_coherent instead of the CMA
API. This also allows it to be built as a module easily.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>