bcm2838.dtsi: Use BCM2711 PCIe compatible string
commit  4d7371c2a8b151f4fd5889c5da5534c784c3bcae from
https://github.com/raspberrypi/linux.git rpi-5.4.y

The BCM2711 PCIe controller has a limited address range in the B0
silicon, and the driver uses a compatible string to identify the
limitation. The current Pi 4 firmware will override the compatible
string if it detects a downstream DTB and it is running on a newer
revision but set the default value to enable the workaround for
backwards-compatibility with old firmware.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>