staging: bcm2835-codec: implement V4L2_CID_MIN_BUFFERS_FOR_CAPTURE
commit  084a65476b6d59b1bce75d037ff253381845c638 from
https://github.com/raspberrypi/linux.git rpi-5.6.y

The stateful decoder specification shows an optional step for retrieving
the miminum number of capture buffers required for the decoder to
proceed. While not a required parameter, having it makes some
applications happy.

bcm2835-codec is a little different from other decoder implementations
in that there is an intermediate format conversion between the hardware
and V4L2 buffers. The number of capture buffers required is therefore
independent of the stream and DPB etc.

There are plans to remove the conversion, but it requires a fair amount
of rework within the firmware. Until that is done, simply return a value
of 1.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>