drm/vc4: Resolve the vblank warnings on mode switching
commit  6bceb44b32a87557b0b439c58a5799a67cfb3c4a from
https://github.com/raspberrypi/linux.git rpi-5.8.y

The details over when and how a driver is to service the
vblank events are sketchy, and the fkms driver was triggering
a kernel warning every time the crtc was enabled or disabled.

Copy the event handling as used by the vc4-kms driver slightly
more closely, and we avoid the warnings.

https://github.com/raspberrypi/linux/issues/3020

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>