usb/dwc2: Set correct state on gadget disconnect
commit  c3974850cc10de2596893eaec2f5c456e9ded6a0 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

When operating as a USB device, when the device is disconnected the suspend interrupt is called rather than the disconnect interrupt, this results in the state remaining as "configured".  This change changes the state to "not attached" instead when the suspend interrupt is called.
Signed-off-by: Meng Li <Meng.Li@windriver.com>