mtd: cadence-quadspi: Added support for DDR PHY mode
This patch adds support for DDR PHY mode.
Controller and flash device will be switched to DDR mode based on the
controller capabilities(8_8_8).
Read Id based RX tuning algorithm is implemented in the driver since
tuning is mandatory for RX to work in DDR PHY mode.
Device Id will be read in NON PHY mode and will be used this as
golden copy for RX tuning.
DDR mode is currently supported only for MICRON devices where as for
other flash devices only NON-PHY mode is supported.

Signed-off-by: Sai Krishna Potthuri <lakshmi.sai.krishna.potthuri@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>