v4l: xilinx: vip: Keep NV16 first in vip table
When Scene change detection IP alone is in the v4l2 pipeline,
the driver is unable to enumerate all the supported pixel formats
based on media code set on the subdev source port. The function
xvip_dma_enum_format() just returns the pix format whichever comes
first in the vip table list.

This patch ensures that NV16 comes first in the vip table. This fixes
the issue temporarily as SCD IP supports only NV16 in 422 8bit
formats.

The permanant fix is to implement a generic interface between V4L2
bridge driver and dma driver to get the supported pixel formats.

Signed-off-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>