ARM64: dts: imx8mm: re-order display nodes
commit  from
https://source.codeaurora.org/external/imx/linux-imx

The devices in the device tree are registered in the
top-down order. But when system suspend entered, the
device tree is walked in a bottom-up order to suspend
devices. So change the display device nodes register
order to be:

LCDIF -> SEC DSIM -> Display Subsystem

Since in display subystem, it will disable the whole
display pipeline. So this should be first suspended
before LCDIF and SEC DSIM. And besides, the SEC DSIM
is better to be suspended before LCDIF which is the
same with the sequence for display pipeline disables.
And when system resume entered, the devices resume
sequence is:

LCDIF -> SEC DSIM -> Display Subsystem

Which is a top-down order and is the correct sequence
for display devices resume.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>