v4l: xilinx: dma: Disable auto restart for low latency capture
Set default mode of operation instead of AUTO_RESTART for low latency
capture as AUTO_RESTART mode may take extra one frame delay
between programming and the actual writing data to memory.

As for the low latency capture we want to submit the buffer to consumer
when dma has just started processing without keeping any extra delay.

And using auto-restart mode without this delay is not recommended as
it may lead to frame tear due to next component trying to use the buffer
while dma is actually writing into it.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>