ASoC: fsl_asrc: Get dma channel from Front-End for p2p
commit  from
https://source.codeaurora.org/external/imx/linux-imx

Previously we get dma channel for p2p from Back-End, but the
channel maybe allocated already by Back-End platform driver.
that we switch to imx-pcm-dma-v2.

But if we get dma channel for p2p from Front-End then we don't
need to switch to imx-pcm-dma-v2 driver. this more formal

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>