zynqmp_r5_remoteproc: Inherit parent device's DMA ops and mask.
When rpmsg/remoteproc is run under Xen, the virtio ring code will attempt
to use the DMA ops for the parent device of the virtqueue device when
running under Xen. Unfortunately, the zynqmp_r5_remoteproc device's
parent device's DMA ops are set to dummy_dma_ops. dummy_dma_ops.map_page
is set to __dummy_map_page(), which just returns 0.

This was observed on zynqmp when the kernel tries to start the firmware
in the R5. zynqmp_r5_probe() tries to set up the rpmsg virtio ring
buffers. It eventually calls vring_map_one_sg(). In normal operation,
this just returns the physical address of the buffer. When running
under Xen, either Dom0 or DomU, vring_map_one_sg() attempts to
use the DMA operations for the parent device to obtain a DMA address
for the buffer. Because the parent's DMA operations were set to
the dummy operations, this resulted in the rpmsg_virtio code trying
to use address 0 as a scatter gather buffer. Usually, this caused
a stack dump, but sometimes it would panic the kernel.

This patch changes the zynqmp_r5_remoteproc driver to inherit the
DMA operations and mask from its parent, which is set up to use
the default ARM architecture operations.

Signed-off-by: Ed Mooring <ed.mooring@xilinx.com>
Acked-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Acked-by: Ben Levinsky <ben.levinsky@xilinx.com>
Tested-by: Ben Levinsky <ben.levinsky@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>