dmaengine: xilinx_dma: Fix SG internal error in cdma prep_dma_sg mode
In CDMA prep_dma_sg mode when source and destination scatterlist length
has more than one element, interrupt is generated on the first segment
(since IRQThreshold=1) instead of complete descriptor. In the driver we
free all descriptor segments on IOC but IP is still working on remaining
descriptor segments and as a result we see an SG internal error. To fix
this behavior set CDMACR.IRQThreshold to numbers of segments in a cdma
descriptor.

Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Acked-by: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>