MLK-22722: ASoC: fsl_easrc: Use global variable instead local variable
commit  from
https://source.codeaurora.org/external/imx/linux-imx

After suspend & resume, there may be noise in asrc output. The reason
is the coeff variable is local, in resume function this variable is
invalid to access, so that cause the prefilter coeff not correctly be
setted.

This patch is to change the coeff variable to be global that we can
get the correct value in resume.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>
(cherry picked from commit 56aa181b23f82280af01dbe1193123d30af61690)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>