staging: vchiq_arm: Unify the unload handling of platform devs
commit  27735dd6dc2fef2c11ce20ac9c8866dcbe8b544d from
https://github.com/raspberrypi/linux.git rpi-5.4.y

A helper function vchiq_register_child was added to deal with
adding the platform devices. This returns NULL on failure, and
that is assigned to the struct platform_device. There is
therefore no way for remove to encounter an error pointer, so
checking for IS_ERR() is redundant.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>