staging: vchiq_arm: Unify the unload handling of platform devs
commit  57ddb294862c8c62a1b5b55e285f9c30e10b34e0 from
https://github.com/raspberrypi/linux.git rpi-5.6.y

A helper function vchiq_register_child was added to deal with
adding the platform devices. This returns NULL on failure, and
that is assigned to the struct platform_device. There is
therefore no way for remove to encounter an error pointer, so
checking for IS_ERR() is redundant.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>