staging: vchiq_arm: Unify the unload handling of platform devs
commit  8aa11a40962812f0504a07b9655c383430751143 from
https://github.com/raspberrypi/linux.git rpi-5.8.y

A helper function vchiq_register_child was added to deal with
adding the platform devices. This returns NULL on failure, and
that is assigned to the struct platform_device. There is
therefore no way for remove to encounter an error pointer, so
checking for IS_ERR() is redundant.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>